+++ src/resources/constants.js
@@ -1,6 +1,6 @@
 export const PIXEL_BANNER_PLUS = {
-    // API_URL: 'https://pixel-banner.online/',
-    API_URL: 'http://localhost:3000/',
+    API_URL: 'https://pixel-banner.online/',
+    // API_URL: 'http://localhost:3000/',
     ENDPOINTS: {
         VERIFY: 'verify',
         GENERATE: 'generate',
@@ -19,7 +19,7 @@ export const PIXEL_BANNER_PLUS = {
         BANNER_VOTES_USER_VOTE: 'api/banner-votes/:id/user-vote',
         BANNER_VOTES_UPVOTE: 'api/banner-votes/:id/upvote',
         BANNER_VOTES_DOWNVOTE: 'api/banner-votes/:id/downvote',
-        VERSION_CHECK: 'version-check'
+        INFO: 'api/pixel-banner-info'
     },
     SHOP_URL: 'https://ko-fi.com/s/7ce609ff2c',
     DONATE_URL: 'https://ko-fi.com/jparkerweb'