+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 All notable changes to the Pixel Banner plugin will be documented in this file.
 
-### v3.1.0 - 2025-03-19
+### v3.1.0 - 2025-03-22
 #### ✨ Added
 - Modal for setting banner source from a URL
 - Add voting functionality to Pixel Banner Store
 - Added clear instruction to the top of the "Set Banner Icon" modal
-- Enhance the visibility of the "🌱 GROW YOUR IDEA" button on the AI Banner modal to make it more prominent, highlighting its usefulness.
+- Enhance the visibility of the "🌱 GROW YOUR IDEA" button on the AI Banner modal to make it more prominent, highlighting its usefulness
+- Show a "Upgrade Available" message in the footer of hte Pixel Banner main menu if a new version is available
 
 #### 📦 Updated
 - Added back ability to prevent the `🚩` select pixel banner icon from being displayed on notes (you can still set the opacity of the icon when enabled)